Fix scheduled advisor import and persistence regressions
